Extract registration/storage helpers into dispatcher/registry.rs
Mechanical move of registration, unregistration, and conflict-check
logic from dispatcher.rs into a dedicated registry submodule. Follows
the established pattern (layers.rs, timeout.rs) of impl Dispatcher
blocks in child modules.

Moved to dispatcher/registry.rs:
- register(), register_binding() — global hotkey registration
- register_sequence(), register_sequence_with_options(),
  register_sequence_binding_with_id(), register_sequence_binding() —
  sequence registration chain
- unregister() — unified binding removal with sequence state cleanup
- is_registered() — hotkey presence query

Moved tests to dispatcher/registry.rs:
- register_returns_unique_id (from dispatcher.rs)
- is_registered_reflects_state (from dispatcher.rs)
- register_sequence_accepts_typed_sequence (from sequence.rs)
- register_sequence_accepts_string_input (from sequence.rs)
- register_sequence_accepts_vec_hotkeys_input (from sequence.rs)
- register_sequence_reports_parse_error_for_string_input (from sequence.rs)
- registering_sequence_with_existing_hotkey_id_is_rejected (from sequence.rs)
- registering_hotkey_with_existing_sequence_id_is_rejected (from sequence.rs)

Cross-cutting tests (e.g. unregister-during-pending-sequence) remain
in their original modules. No behavior changes, no public API changes.
